feat: initialize project structure with core pages, database schema, task worker, and authentication setup
