+++ client/src/App.tsx
@@ -50,7 +50,6 @@ import DailyNotesList from "@/pages/daily-notes";
 import DailyNoteDetail from "@/pages/daily-note-detail";
 import SuperSearchPage from "@/pages/super-search";
 import FamilyTreePage from "@/pages/family-tree";
-import FamilyTreeDevPage from "@/pages/family-tree-dev";
 import NotFound from "@/pages/not-found";
 
 const SEEN_EXPORTS_KEY = "seen_completed_export_task_ids";
@@ -125,7 +124,6 @@ function Router() {
       <ProtectedRoute path="/graph-3d" component={GraphRedirect} />
       <ProtectedRoute path="/social-graph-3d" component={SocialGraph3D} />
       <ProtectedRoute path="/family-tree" component={FamilyTreePage} />
-      <ProtectedRoute path="/family-tree-dev-version" component={FamilyTreeDevPage} />
       <ProtectedRoute path="/elo-ranking" component={EloRanking} />
       <ProtectedRoute path="/guess-the-sex" component={GuessTheSex} />
       <ProtectedRoute path="/account-matching" component={AccountMatching} />
